requirements.txt
requests==2.32.3
bs4==0.0.2
larksuite-oapi==1.0.34
google-api-python-client==2.169.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gspread==6.2.0